backend/requirements.txt
sentence-transformers
langchain
huggingface-hub
langchain-huggingface
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain-google-genai
langchain_experimental
-e .

